Have SaltedOutPointHasher return a uint64_t
We no longer use a boost map therefore we can return a uin64_t
on all systems rather than size_t.  This was previously an issue
on a narrow range of boost versions which resulted in failures when syncing
the chain (Core issue #4634).

+++ src/coins.h
@@ -111,12 +111,7 @@ private:
 public:
     SaltedOutpointHasher();
 
-    /**
-     * This *must* return size_t. With Boost 1.46 on 32-bit systems the
-     * unordered_map will behave unpredictably if the custom hasher returns a
-     * uint64_t, resulting in failures when syncing the chain (#4634).
-     */
-    size_t operator()(const COutPoint &id) const { return SipHashUint256Extra(k0, k1, id.hash, id.n); }
+    uint64_t operator()(const COutPoint& id) const { return SipHashUint256Extra(k0, k1, id.hash, id.n); }
 };
 
 struct CCoinsCacheEntry
